attach: rename queue library to prestore
