+++ projects/rocprofiler-sdk/source/lib/rocprofiler-sdk-prestore/queue_registration.cpp
